course UI 작업
